CSFix, add new flash data when category not found

+++ src/Controller/CategoryController.php
@@ -7,7 +7,6 @@ use App\Repository\CategoryRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 
 class CategoryController extends AbstractController
@@ -16,7 +15,7 @@ class CategoryController extends AbstractController
      * @Route("/categories", name="categories")
      *
      * @param \Symfony\Component\HttpFoundation\Request $request
-     * @param \App\Repository\CategoryRepository $categoryRepository
+     * @param \App\Repository\CategoryRepository        $categoryRepository
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
@@ -39,30 +38,33 @@ class CategoryController extends AbstractController
 
         return $this->render('category/index.html.twig', [
             'form' => $form->createView(),
-            'categories' => $categoryRepository->findAll()
+            'categories' => $categoryRepository->findAll(),
         ]);
     }
 
     /**
      * @Route("/categories/{categoryId}/delete", name="categories.destroy", methods={"DELETE"})
      *
-     * @param int $categoryId
+     * @param int                                $categoryId
      * @param \App\Repository\CategoryRepository $categoryRepository
-     * @param \Symfony\Component\HttpFoundation\Session\SessionInterface $session
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function destroy(int $categoryId, CategoryRepository $categoryRepository, SessionInterface $session): Response
+    public function destroy(int $categoryId, CategoryRepository $categoryRepository): Response
     {
         $category = $categoryRepository->find($categoryId);
 
-        $em = $this->getDoctrine()->getManager();
+        if (null !== $category) {
+            $em = $this->getDoctrine()->getManager();
 
-        $em->remove($category);
+            $em->remove($category);
 
-        $em->flush();
+            $em->flush();
 
-        $this->addFlash('category.deletion', sprintf('Category %s deleted', $category->getName()));
+            $this->addFlash('category.deletion.success', sprintf('Category %s deleted', $category->getName()));
+        } else {
+            $this->addFlash('category.deletion.error', 'Category not found');
+        }
 
         return $this->redirectToRoute('categories');
     }
